backend/requirements.txt
fastapi
uvicorn
ortools
python-dotenv
pandas
geopy
google-generativeai
python-dotenv
